data_structures/base/day07/word_pattern_test.go
package day07

import "testing"

// TestWordPattern 测试单词规律
func TestWordPattern(t *testing.T) {
	type args struct {
		pattern string
		s       string
	}
	tests := []struct {
		name string
		args args
		want bool
	}{
		{
			name: "01",
			args: args{
				pattern: "abba",
				s:       "dog cat cat dog",
			},
			want: true,
		},
		{
			name: "02",
			args: args{
				pattern: "abba",
				s:       "dog cat cat fish",
			},
			want: false,
		},
		{
			name: "03",
			args: args{
				pattern: "aaaa",
				s:       "dog cat cat dog",
			},
			want: false,
		},
		{
			name: "04",
			args: args{
				pattern: "abba",
				s:       "dog dog dog dog",
			},
			want: false,
		},
		{
			name: "05",
			args: args{
				pattern: "aba",
				s:       "dog dog dog dog",
			},
			want: false,
		},
	}
	for _, tt := range tests {
		t.Run(
			tt.name, func(t *testing.T) {
				if got := wordPattern(tt.args.pattern, tt.args.s); got != tt.want {
					t.Errorf("wordPattern() = %v, want %v", got, tt.want)
				}
			},
		)
	}
}
